@@ -7,14 +7,14 @@ homepage: https://github.com/rodion-m/dio_smart_retry
 documentation: https://github.com/rodion-m/dio_smart_retry#contents
 
 environment:
-  sdk: '>=2.17.0 <3.0.0'
+  sdk: '>=3.0.0 <4.0.0'
 
 dependencies:
-  dio: ^5.1.1
+  dio: ^5.3.3
   http_parser: ^4.0.2
   path: ^1.8.3
 
 dev_dependencies:
-  dart_code_metrics: ^5.7.2
-  test: ^1.24.2
-  very_good_analysis: ^4.0.0+1
+  dart_code_metrics: ^5.7.6
+  test: ^1.24.9
+  very_good_analysis: ^5.1.0
