@@ -6,7 +6,10 @@ import 'package:dio_smart_retry/src/http_status_codes.dart';
 import 'package:dio_smart_retry/src/multipart_file_recreatable.dart';
 import 'package:dio_smart_retry/src/retry_not_supported_exception.dart';
 
-typedef RetryEvaluator = FutureOr<bool> Function(DioError error, int attempt);
+typedef RetryEvaluator = FutureOr<bool> Function(
+  DioException error,
+  int attempt,
+);
 
 /// An interceptor that will try to send failed request again
 class RetryInterceptor extends Interceptor {
@@ -82,11 +85,11 @@ class RetryInterceptor extends Interceptor {
 
   /// Redirects to [DefaultRetryEvaluator.evaluate]
   ///   with [defaultRetryableStatuses]
-  static final FutureOr<bool> Function(DioError error, int attempt)
+  static final FutureOr<bool> Function(DioException error, int attempt)
       defaultRetryEvaluator =
       DefaultRetryEvaluator(defaultRetryableStatuses).evaluate;
 
-  Future<bool> _shouldRetry(DioError error, int attempt) async {
+  Future<bool> _shouldRetry(DioException error, int attempt) async {
     try {
       return await _retryEvaluator(error, attempt);
     } catch (e) {
@@ -105,7 +108,10 @@ class RetryInterceptor extends Interceptor {
   }
 
   @override
-  Future<dynamic> onError(DioError err, ErrorInterceptorHandler handler) async {
+  Future<dynamic> onError(
+    DioException err,
+    ErrorInterceptorHandler handler,
+  ) async {
     if (err.requestOptions.disableRetry) {
       return super.onError(err, handler);
     }
@@ -135,7 +141,7 @@ class RetryInterceptor extends Interceptor {
         requestOptions = _recreateOptions(err.requestOptions);
       } on RetryNotSupportedException catch (e) {
         return super.onError(
-          DioError(requestOptions: requestOptions, error: e),
+          DioException(requestOptions: requestOptions, error: e),
           handler,
         );
       }
@@ -153,7 +159,7 @@ class RetryInterceptor extends Interceptor {
       await dio
           .fetch<void>(requestOptions)
           .then((value) => handler.resolve(value));
-    } on DioError catch (e) {
+    } on DioException catch (e) {
       super.onError(e, handler);
     }
   }
@@ -191,6 +197,7 @@ class RetryInterceptor extends Interceptor {
   }
 
   var _multipartFileChecked = false;
+
   void _printErrorIfRequestHasMultipartFile(RequestOptions options) {
     if (_multipartFileChecked) return;
     if (options.data is FormData) {
