@@ -20,8 +20,8 @@ void main() {
 
     try {
       await dio.get<dynamic>('https://mock.codes/400');
-    } on DioError catch (error) {
-      if (error.type != DioErrorType.badResponse ||
+    } on DioException catch (error) {
+      if (error.type != DioExceptionType.badResponse ||
           error.response?.statusCode != 400) {
         rethrow;
       }
